extract legacy v1 code into separate module
